app/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
requests>=2.31
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
pypdf2>=3.0.1
lxml>=5.0.0
blingfire>=0.1.8
pysbd>=0.3.4
